Update fullscreen mode with redux
