+++ src/redux/player/actionTypes.js
@@ -1,1 +1,2 @@
 export const SET_PLAY_STATUS = "SET_PLAY_STATUS"
+export const SET_FULL_SCREEN_STATUS = "SET_FULL_SCREEN_STATUS"
