+++ src/redux/player/reducer.js
@@ -1,9 +1,11 @@
 import { 
     SET_PLAY_STATUS,
+    SET_FULL_SCREEN_STATUS
 } from './actionTypes';
 
 const initialState = {
     isPlaying: false,
+    isFullScreen: false,
 }
 
 const reducer = (state = initialState, {type, payload}) => {
@@ -13,6 +15,11 @@ const reducer = (state = initialState, {type, payload}) => {
             return {
                 ...state,
                 isPlaying: payload
+            }
+        case SET_FULL_SCREEN_STATUS:
+            return {
+                ...state,
+                isFullScreen: payload
             }
         default:
             return state
